% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HellenicLandings.R
\docType{data}
\name{HellenicLandings}
\alias{HellenicLandings}
\title{Selected Hellenic Fishery Landings}
\format{A data frame with variables:
\describe{
  \item{Year}{}
  \item{Anchovy}{}
  \item{Pilchard}{}
  \item{Chub.mackerel}{}
  \item{Horse.mackerel}{}
  \item{Mackerel}{}
  \item{Jack.mackerel}{}
}}
\usage{
HellenicLandings
}
\description{
Annual total landings (in metric tons) in Hellenic (Greek) waters 1964-2007 from the Hellenic Statisitcal Authority, Digital Library (ELSTAT), Special Publications, Fisheries.
}
\details{
Reports are at http://dlib.statistics.gr/portal/page/portal/ESYE/categoryyears?p_cat=10007977&p_topic=10007977 .  The data are from the Total column in Table IV in the publications.
}
\keyword{datasets}
