% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNOTEL.R
\docType{data}
\name{SNOTEL}
\alias{SNOTEL}
\title{SNOTEL Data}
\format{A data frame with variables:
\describe{
  \item{Station.Id}{}
  \item{Station.Name}{}
  \item{Year}{}
  \item{Jan..Dec}{}
}}
\usage{
SNOTEL
}
\description{
Snow measurement data from SNOTEL stations in the United States.
}
\details{
Data downloaded from the National Weather and Climate Center.
}
\keyword{datasets}
