% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNOTEL.R
\docType{data}
\name{SNOTELmeta}
\alias{SNOTELmeta}
\title{SNOTEL Station Data}
\format{A data frame with variables:
\describe{
  \item{Station_Name}{}
  \item{Station.Id}{}
  \item{State.Code}{}
  \item{Network.Code}{}
  \item{Network.Name}{}
  \item{Elevation}{}
  \item{Latitude}{}
  \item{Longitude}{}
}}
\usage{
SNOTELmeta
}
\description{
Meta data on each SNOTEL station.
}
\details{
Data downloaded from the National Weather and Climate Center.
}
\keyword{datasets}
